Get dir attribute integration tests (#1220)
* adding test for dir attribute

* renaming

* fixing bytes comparison

* adding more tests

* adding sub dir attribute tests

* fix failure

* adding comment

* adding const variable

* removing unnecessary file

* fixing comment

* fix lint
